+++ src/main/kotlin/org/jetbrains/intellij/IntelliJPluginExtension.kt
@@ -18,7 +18,7 @@ import org.jetbrains.intellij.utils.DependenciesDownloader
 import javax.inject.Inject
 
 /**
- * After the Gradle IntelliJ Plugin is applied, the `intellij` extension can be used to configure the plugin and common settings of the provided tasks.
+ * After the IntelliJ Platform Gradle Plugin is applied, the `intellij` extension can be used to configure the plugin and common settings of the provided tasks.
  *
  * It is mandatory to specify at least the [version] property.
  */