+++ src/main/kotlin/org/jetbrains/intellij/tasks/SetupDependenciesTask.kt
@@ -17,7 +17,7 @@ import org.jetbrains.intellij.logCategory
  *
  * This task is automatically added to the ["After Sync" Gradle trigger](https://www.jetbrains.com/help/idea/work-with-gradle-tasks.html#config_triggers_gradle) to make the IntelliJ SDK dependency available for IntelliJ IDEA right after the Gradle synchronization.
  *
- * > After removing the Gradle IntelliJ Plugin from your project, the `Task 'setupDependencies' not found in root project` exception may occur.
+ * > After removing the IntelliJ Platform Gradle Plugin from your project, the `Task 'setupDependencies' not found in root project` exception may occur.
  *
  * @see <a href="https://plugins.jetbrains.com/docs/intellij/tools-gradle-intellij-plugin-faq.html#task-setupdependencies-not-found-in-root-project">Frequently Asked Questions</a>
  */